Move GCP logging setup to own file
