+++ docker/exporter/exporter.py
@@ -23,7 +23,6 @@ from typing import List
 
 from google.cloud import ndb
 from google.cloud import storage
-from google.cloud import logging as google_logging
 
 import osv
 
@@ -32,46 +31,6 @@ DEFAULT_WORK_DIR = '/work'
 DEFAULT_EXPORT_BUCKET = 'osv-vulnerabilities'
 _EXPORT_WORKERS = 32
 ECOSYSTEMS_FILE = 'ecosystems.txt'
-
-
-def _setup_gcp_logging():
-  """Set up GCP logging and error reporting."""
-
-  logging_client = google_logging.Client()
-  logging_client.setup_logging()
-
-  old_factory = logging.getLogRecordFactory()
-
-  def record_factory(*args, **kwargs):
-    """Insert jsonPayload fields to all logs."""
-
-    record = old_factory(*args, **kwargs)
-    if not hasattr(record, 'json_fields'):
-      record.json_fields = {}
-
-    # Add jsonPayload fields to logs that don't contain stack traces to enable
-    # capturing and grouping by error reporting.
-    # https://cloud.google.com/error-reporting/docs/formatting-error-messages#log-text
-    if record.levelno >= logging.ERROR and not record.exc_info:
-      record.json_fields.update({
-          '@type':
-              'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',  # pylint: disable=line-too-long
-          'serviceContext': {
-              'service': 'exporter',
-          },
-          'context': {
-              'reportLocation': {
-                  'filePath': record.pathname,
-                  'lineNumber': record.lineno,
-                  'functionName': record.funcName,
-              }
-          },
-      })
-
-    return record
-
-  logging.setLogRecordFactory(record_factory)
-  logging.getLogger().setLevel(logging.INFO)
 
 
 class Exporter:
@@ -167,6 +126,6 @@ def main():
 
 if __name__ == '__main__':
   _ndb_client = ndb.Client()
-  _setup_gcp_logging()
+  osv.logging_helper.setup_gcp_logging('exporter')
   with _ndb_client.context():
     main()
