+++ docker/importer/importer.py
@@ -25,7 +25,6 @@ from typing import List, Tuple, Optional
 from google.cloud import ndb
 from google.cloud import pubsub_v1
 from google.cloud import storage
-from google.cloud import logging as google_logging
 import pygit2
 
 import osv
@@ -41,49 +40,6 @@ _NO_UPDATE_MARKER = 'OSV-NO-UPDATE'
 _BUCKET_THREAD_COUNT = 20
 
 _client_store = threading.local()
-
-
-def _setup_gcp_logging():
-  """Set up GCP logging and error reporting."""
-
-  logging_client = google_logging.Client()
-  logging_client.setup_logging()
-
-  old_factory = logging.getLogRecordFactory()
-
-  def record_factory(*args, **kwargs):
-    """Insert jsonPayload fields to all logs."""
-
-    record = old_factory(*args, **kwargs)
-    if not hasattr(record, 'json_fields'):
-      record.json_fields = {}
-
-    # Add jsonPayload fields to logs that don't contain stack traces to enable
-    # capturing and grouping by error reporting.
-    # https://cloud.google.com/error-reporting/docs/formatting-error-messages#log-text
-    if record.levelno >= logging.ERROR and not record.exc_info:
-      record.json_fields.update({
-          '@type':
-              'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',  # pylint: disable=line-too-long
-          'serviceContext': {
-              'service': 'importer',
-          },
-          'context': {
-              'reportLocation': {
-                  'filePath': record.pathname,
-                  'lineNumber': record.lineno,
-                  'functionName': record.funcName,
-              }
-          },
-      })
-
-    return record
-
-  logging.setLogRecordFactory(record_factory)
-  logging.getLogger().setLevel(logging.INFO)
-  logging.getLogger('google.api_core.bidi').setLevel(logging.ERROR)
-  logging.getLogger('google.cloud.pubsub_v1.subscriber._protocol.'
-                    'streaming_pull_manager').setLevel(logging.ERROR)
 
 
 def _is_vulnerability_file(source_repo, file_path):
@@ -542,7 +498,7 @@ def main():
 
 
 if __name__ == '__main__':
-  _setup_gcp_logging()
+  osv.logging_helper.setup_gcp_logging('importer')
   _ndb_client = ndb.Client()
   with _ndb_client.context():
     main()
