+++ docker/worker/Pipfile
@@ -6,7 +6,6 @@ name = "pypi"
 [packages]
 jsonschema = "*"
 google-cloud-pubsub = "*"
-google-cloud-logging = "*"
 google-cloud-ndb = "*"
 PyYAML = "*"
 google-cloud-storage = "*"
