+++ docker/worker/worker.py
@@ -31,7 +31,6 @@ import google.cloud.exceptions
 from google.cloud import ndb
 from google.cloud import pubsub_v1
 from google.cloud import storage
-from google.cloud import logging as google_logging
 
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 import osv
@@ -88,11 +87,8 @@ class UpdateConflictError(Exception):
   """Update conflict exception."""
 
 
-def _setup_gcp_logging():
-  """Set up GCP logging and error reporting."""
-
-  logging_client = google_logging.Client()
-  logging_client.setup_logging()
+def _setup_logging_extra_info():
+  """Set up extra GCP logging information."""
 
   old_factory = logging.getLogRecordFactory()
 
@@ -107,32 +103,9 @@ def _setup_gcp_logging():
       record.json_fields['source_id'] = _state.source_id
     record.json_fields['thread'] = record.thread
 
-    # Add jsonPayload fields to logs that don't contain stack traces to enable
-    # capturing and grouping by error reporting.
-    # https://cloud.google.com/error-reporting/docs/formatting-error-messages#log-text
-    if record.levelno >= logging.ERROR and not record.exc_info:
-      record.json_fields.update({
-          '@type':
-              'type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent',  # pylint: disable=line-too-long
-          'serviceContext': {
-              'service': 'worker',
-          },
-          'context': {
-              'reportLocation': {
-                  'filePath': record.pathname,
-                  'lineNumber': record.lineno,
-                  'functionName': record.funcName,
-              }
-          },
-      })
-
     return record
 
   logging.setLogRecordFactory(record_factory)
-  logging.getLogger().setLevel(logging.INFO)
-  logging.getLogger('google.api_core.bidi').setLevel(logging.ERROR)
-  logging.getLogger('google.cloud.pubsub_v1.subscriber._protocol.'
-                    'streaming_pull_manager').setLevel(logging.ERROR)
 
 
 class _PubSubLeaserThread(threading.Thread):
@@ -709,5 +682,6 @@ def main():
 
 
 if __name__ == '__main__':
-  _setup_gcp_logging()
+  osv.logging_helper.setup_gcp_logging('worker')
+  _setup_logging_extra_info()
   main()
